refactored to prepare for dependency injections implementation
